% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv.r
\name{is.rv}
\alias{is.rv}
\title{Check if an object is a discrete random variable.}
\usage{
is.rv(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if an object is a discrete random variable.
}
\examples{
is.rv(1:10)
is.rv(rv(1:10))
}

