% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{E}
\alias{SD}
\alias{VAR}
\alias{moments}
\title{Compute the moments of a random variable.}
\usage{
E(x)

VAR(x)

SD(x)
}
\arguments{
\item{x}{random variable to summarise}
}
\description{
Compute the moments of a random variable.
}
\examples{
dice <- rv(1:6)
E(dice)
VAR(dice)
SD(dice)
}
\seealso{
\code{\link{Z}} to standardise an rv by subtracting off mean
  and dividing by standard deviation
}

